+++ lumina.out.js
@@ -1,5 +1,9 @@
 import { io, str, math, list, Result, Option, __set, formatValue, LuminaPanic } from "./lumina-runtime.js";
 function main() {
-  __unknown__;
+  while (true) {
+    let count_1 = 1;
+  }
+  let count_2 = count_1;
+  count_2;
 }
 export { io, str, math, list, Result, Option, __set, formatValue, LuminaPanic };